\name{concerto-package}
\alias{concerto-package}
\docType{package}
\title{
Package for Concerto platform test logic R code.
}
\description{
Contains all required functions by Concerto platform. 
}
\details{
\tabular{ll}{
Package: \tab concerto\cr
Type: \tab Package\cr
Version: \tab 0.3.6\cr
Date: \tab 2014-06-18\cr
License: \tab GNU GPL v2\cr
}

}
\author{
Przemyslaw Lis

Maintainer: Przemyslaw Lis <pl362@cam.ac.uk>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
