\name{concerto.db.connect}
\alias{concerto.db.connect}
\title{
MySQL database connection.
}
\description{
This function establishes connection with MySQL database.
}
\usage{
concerto.db.connect(user, password, dbName, host = "localhost", port = 3306, dbTimezone)
}
\arguments{
  \item{user}{
MySQL user login.
}
  \item{password}{
MySQL user password.
}
  \item{dbName}{
MySQL database name.
}
  \item{host}{
MySQL server host.
}
  \item{port}{
MySQL server port.
}
  \item{dbTimezone}{
MySQL timezone to use.
}
}
\details{
}
\value{
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
}
\keyword{ MySQL }
\keyword{ database }
