% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.finalize.R
\name{concerto.finalize}
\alias{concerto.finalize}
\title{concerto.finalize}
\usage{
concerto.finalize()
}
\description{
Undocumented function used to update status and close DB connection
}
\author{
Przemyslaw Lis
}

