\name{concerto.interpretResponse}
\alias{concerto.interpretResponse}
\title{
Test participant interaction.
}
\description{
Interprets test participant response.
}
\usage{
concerto.interpretResponse()
}
\details{
}
\value{
Decoded JSON list with test particpant response.
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
}
\keyword{ session }
