\name{concerto.qti.contains}
\alias{concerto.qti.contains}
\title{
Concerto QTI helper function
}
\description{
Checks if second expression result is contained in first expression result, given the specified variable base type and cardinality.
}
\usage{
concerto.qti.contains(exp1, exp2, baseType, cardinality)
}
\arguments{
  \item{exp1}{
Expression that is going to be searched.
}
  \item{exp2}{
Expression that we are going to search for.
}
  \item{baseType}{
Base type of expressions results.
}
  \item{cardinality}{
Cardinality of expression.
}
}
\details{
}
\value{
Logical: TRUE if second expression is found or FALSE otherwise.
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
}

\keyword{ QTI }
