\name{concerto.qti.delete}
\alias{concerto.qti.delete}
\title{
Concerto QTI helper function
}
\description{
Removes object elements.
}
\usage{
concerto.qti.delete(exp1, exp2, baseType)
}
\arguments{
  \item{exp1}{
Elements that we want to remove.
}
  \item{exp2}{
Object to remove elements from.
}
  \item{baseType}{
Base type of object.
}
}
\details{
}
\value{
Object with elements removed.
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
}

\keyword{ QTI }