\name{concerto.qti.equal}
\alias{concerto.qti.equal}
\title{
Concerto QTI helper function
}
\description{
Checks if first object is equal to the second one.
}
\usage{
concerto.qti.equal(arg1, arg2, baseType)
}
\arguments{
  \item{arg1}{
First object.
}
  \item{arg2}{
Second object.
}
  \item{baseType}{
Objects base type.
}
}
\details{
}
\value{
Logical: TRUE if objects are equal and FALSE otherwise.
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
}

\keyword{ QTI }
