\name{concerto.qti.initialize}
\alias{concerto.qti.initialize}
\title{
QTI initialization
}
\description{
Initializes QTI object.
}
\usage{
concerto.qti.initialize(qtiID, params = list(), workspaceID = concerto$workspaceID)
}
\arguments{
  \item{qtiID}{
QTI object ID or name.
}
  \item{params}{
Named list with parameters we want to pass to QTI initialization.
}
  \item{workspaceID}{
Concerto workspace ID where QTI object is located.
}
}
\details{
}
\value{
Named list with QTI session state and:
\item{QTI_HTML }{HTML generated by QTI logic}
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
init <- concerto.qti.initialize(1)
}

\keyword{ QTI }
