\name{concerto.qti.mapResponse}
\alias{concerto.qti.mapResponse}
\title{
Concerto QTI helper function
}
\description{
Equivalent of QTI mapResponse expression.
}
\usage{
concerto.qti.mapResponse(variableName)
}
\arguments{
  \item{variableName}{
Variable to which response should be mapped.
}
}
\details{
}
\value{
Response mapping result.
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{

}

\keyword{ QTI }