\name{concerto}
\alias{concerto}
\alias{concerto$testID}
\alias{concerto$sessionID}
\alias{concerto$workspaceID}
\alias{concerto$workspacePrefix}
\alias{concerto$templateFIFOPath}
\alias{concerto$sessionPath}
\alias{concerto$mediaPath}
\alias{concerto$userIP}
\alias{concerto$db$connection}
\alias{concerto$db$name}
\docType{data}
\title{
List of Concerto variables.
}
\description{
List of Concerto variables containing: testID, sessionID, workspaceID, workspacePrefix, templateFIFOPath, sessionPath, mediaPath, userIP, db$connection and db$name.
}
\usage{
concerto$testID
concerto$sessionID
concerto$workspaceID,
concerto$workspacePrefix
concerto$templateFIFOPath
concerto$sessionPath
concerto$mediaPath
concerto$userIP
concerto$db$connection
concerto$db$name
}
\format{
}
\details{
}
\source{
}
\references{
}
\examples{
print(concerto$sessionID)
}
\keyword{concerto}
