\name{concerto.table.query}
\alias{concerto.table.query}
\title{
Table query
}
\description{
Perform query on Concerto table.
}
\usage{
concerto.table.query(sql)
}
\arguments{
  \item{sql}{
SQL query to perform.
}
}
\details{
Contains extended function wizard with SQL query builder when used in Concerto platform.
}
\value{
Returns result set.
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
}

\keyword{ table }
\keyword{ MySQL }
\keyword{ database }
