% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.test.get.R
\name{concerto.test.get}
\alias{concerto,}
\alias{concerto.test.get}
\alias{test.get}
\title{Test object retrieval}
\usage{
concerto.test.get(testID, workspaceID = concerto$workspaceID)
}
\arguments{
\item{testID}{Test ID or name.}

\item{workspaceID}{Concerto workspace ID where test object is located.}
}
\value{
List with test object properties:
\itemize{
\item{id}{test object ID}
\item{name}{test object name}
\item{code}{test logic R code}
}
}
\description{
Gets template object.
}
\examples{
\dontrun{
test1 <- concerto.test.get(1)
test2 <- concerto.test.get("some_other_test")
}
}
\author{
Przemyslaw Lis
}

