% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.test.run.R
\name{concerto.test.run}
\alias{concerto.test.run}
\title{Running Concerto test}
\usage{
concerto.test.run(testID, params = list(),
  workspaceID = concerto$workspaceID)
}
\arguments{
\item{testID}{Test ID or name.}

\item{params}{Named list of parameters you want to pass to the test logic.}

\item{workspaceID}{Concerto workspace ID where the test object is located.}
}
\value{
Named list with the test specific return variables.
}
\description{
Runs Concerto test.
}
\examples{
\dontrun{
test <- concerto.test.run(1)
}
}
\author{
Przemyslaw Lis
}

