\name{concerto.test.updateAllReturnVariables}
\alias{concerto.test.updateAllReturnVariables}
\title{
Updating test returned variables
}
\description{
Updates test all return variables.
}
\usage{
concerto.test.updateAllReturnVariables()
}
\details{
}
\value{
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
concerto.test.updateAllReturnedVariables()
}

\keyword{ test }
