\name{concerto.test.updateReturnVariable}
\alias{concerto.test.updateReturnVariable}
\title{
Updating test return variable
}
\description{
Updates specified test return variable.
}
\usage{
concerto.test.updateReturnVariable(variable)
}
\arguments{
  \item{variable}{
Test return variable name we want to be updated.
}
}
\details{
}
\value{
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
concerto.test.updateReturnVariable("var1")
}
\keyword{ test }
