% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateEffectHide.R
\name{concerto.updateEffectHide}
\alias{concerto.updateEffectHide}
\title{concerto.updateEffectHide}
\usage{
concerto.updateEffectHide(effectHide)
}
\arguments{
\item{effectHide}{Not clearly documented}
}
\description{
Not clearly documented
}
\author{
Przemyslaw Lis
}

