% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateEffectHideOptions.R
\name{concerto.updateEffectHideOptions}
\alias{concerto.updateEffectHideOptions}
\title{concerto.updateEffectHideOptions}
\usage{
concerto.updateEffectHideOptions(effectHideOptions)
}
\arguments{
\item{effectHideOptions}{Not clearly documented}
}
\description{
Not clearly documented
}
\author{
Przemyslaw Lis
}

