% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateEffectShowOptions.R
\name{concerto.updateEffectShowOptions}
\alias{concerto.updateEffectShowOptions}
\title{concerto.updateEffectShowOptions}
\usage{
concerto.updateEffectShowOptions(effectShowOptions)
}
\arguments{
\item{effectShowOptions}{Not clearly documented}
}
\description{
Not clearly documented
}
\author{
Przemyslaw Lis
}

