\name{concerto.updateHead}
\alias{concerto.updateHead}
\title{
Update test session data.
}
\description{
Updates test session head data.
}
\usage{
concerto.updateHead(html)
}
\arguments{
  \item{html}{
Test session head tag code.
}
}
\details{
}
\value{
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
}

\keyword{ session }
