\name{concerto.updateHTML}
\alias{concerto.updateHTML}
\title{
Update test session data.
}
\description{
Updates test session HTML data.
}
\usage{
concerto.updateHTML(html)
}
\arguments{
  \item{html}{
Test session current HTML code.
}
}
\details{
}
\value{
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
}

\keyword{ session }
