% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateLoaderEffectHide.R
\name{concerto.updateLoaderEffectHide}
\alias{concerto.updateLoaderEffectHide}
\title{concerto.updateLoaderEffectHide}
\usage{
concerto.updateLoaderEffectHide(effectHide)
}
\arguments{
\item{effectHide}{Not clearly documented}
}
\description{
Not clearly documented
}
\author{
Przemyslaw Lis
}

