% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateLoaderEffectHideOptions.R
\name{concerto.updateLoaderEffectHideOptions}
\alias{concerto.updateLoaderEffectHideOptions}
\title{concerto.updateLoaderEffectHideOptions}
\usage{
concerto.updateLoaderEffectHideOptions(effectHideOptions)
}
\arguments{
\item{effectHideOptions}{Not clearly documented}
}
\description{
Not clearly documented
}
\author{
Przemyslaw Lis
}

