% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateLoaderEffectShow.R
\name{concerto.updateLoaderEffectShow}
\alias{concerto.updateLoaderEffectShow}
\title{concerto.updateLoaderEffectShow}
\usage{
concerto.updateLoaderEffectShow(effectShow)
}
\arguments{
\item{effectShow}{Not clearly documented}
}
\description{
Not clearly documented
}
\author{
Przemyslaw Lis
}

