% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateLoaderEffectShowOptions.R
\name{concerto.updateLoaderEffectShowOptions}
\alias{concerto.updateLoaderEffectShowOptions}
\title{concerto.updateLoaderEffectShowOptions}
\usage{
concerto.updateLoaderEffectShowOptions(effectShowOptions)
}
\arguments{
\item{effectShowOptions}{Not clearly documented}
}
\description{
Not clearly documented
}
\author{
Przemyslaw Lis
}

