% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateLoaderHead.R
\name{concerto.updateLoaderHead}
\alias{concerto.updateLoaderHead}
\title{concerto.updateLoaderHead}
\usage{
concerto.updateLoaderHead(loaderHead)
}
\arguments{
\item{loaderHead}{Not clearly documented}
}
\description{
Not clearly documented
}
\author{
Przemyslaw Lis
}

