% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateLoaderHTML.R
\name{concerto.updateLoaderHTML}
\alias{concerto.updateLoaderHTML}
\title{concerto.updateLoaderHTML}
\usage{
concerto.updateLoaderHTML(loaderHTML)
}
\arguments{
\item{loaderHTML}{Not clearly documented}
}
\description{
Not clearly documented
}
\author{
Przemyslaw Lis
}

