% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateLoaderTemplateWorkspaceID.R
\name{concerto.updateLoaderTemplateWorkspaceID}
\alias{concerto.updateLoaderTemplateWorkspaceID}
\title{concerto.updateLoaderTemplateWorkspaceID}
\usage{
concerto.updateLoaderTemplateWorkspaceID(workspaceID)
}
\arguments{
\item{workspaceID}{ID field for the workspace}
}
\description{
Set method for loader template workspace ID object
}
\author{
Przemyslaw Lis
}

