\name{concerto.updateRelease}
\alias{concerto.updateRelease}
\title{
Update test session data.
}
\description{
Updates test session release data.
}
\usage{
concerto.updateRelease(release)
}
\arguments{
  \item{release}{
numeric: 0 if test session resources should be released and 1 if they shouldn't
}
}
\details{
}
\value{
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
}

\keyword{ session }
