% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateState.R
\name{concerto.updateState}
\alias{concerto.updateState}
\title{concerto.updateState}
\usage{
concerto.updateState()
}
\description{
Set method for test session state
}
\author{
Przemyslaw Lis
}

