\name{concerto.updateStatus}
\alias{concerto.updateStatus}
\title{
Update test session data.
}
\description{
Updates test session status data.
}
\usage{
concerto.updateStatus(status)
}
\arguments{
  \item{status}{
Test session status.
}
}
\details{
}
\value{
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
}

\keyword{ session }
