\name{concerto.updateTemplateID}
\alias{concerto.updateTemplateID}
\title{
Update test session data.
}
\description{
Updates test session template ID.
}
\usage{
concerto.updateTemplateID(templateID)
}
\arguments{
  \item{templateID}{
Template ID.
}
}
\details{
}
\value{
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
}

\keyword{ session }
