\name{concerto.updateTemplateWorkspaceID}
\alias{concerto.updateTemplateWorkspaceID}
\title{
Update test session data.
}
\description{
Updates test session template workspace ID.
}
\usage{
concerto.updateTemplateWorkspaceID(workspaceID)
}
\arguments{
  \item{workspaceID}{
Concerto workspace ID where template object is located.
}
}
\details{
}
\value{
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
}

\keyword{ session }
