\name{concerto.updateTimeLimit}
\alias{concerto.updateTimeLimit}
\title{
Update test session data.
}
\description{
Updates test session time limit.
}
\usage{
concerto.updateTimeLimit(timeLimit)
}
\arguments{
  \item{timeLimit}{
Test session current template time limit.
}
}
\details{
}
\value{
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
}

\keyword{ session }
