% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.workspace.get.R
\name{concerto.workspace.get}
\alias{concerto.workspace.get}
\title{concerto.workspace.get}
\usage{
concerto.workspace.get(workspaceID)
}
\arguments{
\item{workspaceID}{ID field for the workspace}
}
\value{
Returns workspace name string
}
\description{
Undocumented function
}
\author{
Przemyslaw Lis
}

