% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.R
\docType{data}
\name{config}
\alias{config}
\title{An R Data object containing sample configuration files}
\format{List of character vectors}
\usage{
data(config)
}
\description{
A list object storing a copy of my php.ini and SETTINGS.php files
}

