% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.template.fillHTML.R
\name{template}
\alias{concerto.template.fillHTML}
\alias{template}
\title{Template helper function.}
\usage{
concerto.template.fillHTML(html, params = list())
}
\arguments{
\item{html}{HTML template.}

\item{params}{Named list with parameters.}
}
\value{
HTML with inserted parameters.
}
\description{
Fills template HTML with given parameters.
}
\examples{
\dontrun{
html <- "<html><body>Inserted variable:   param1  </body></html>"
html <- concerto.template.fillHTML(html, list(param1 = "value"))
}
}
\author{
Przemyslaw Lis
}

