% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.R
\docType{package}
\name{concerto-package}
\alias{concerto}
\alias{concerto-package}
\title{Package for Concerto platform test logic R code.}
\description{
Contains all required functions by Concerto platform.
}
\details{
\tabular{ll}{
Package: \tab concerto\cr
Type: \tab Package\cr
Version: \tab 0.3.3\cr
Date: \tab 2013-06-17\cr
License: \tab GNU GPL v2\cr
}
}

