% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.containsOrderedVector.R
\name{concerto.containsOrderedVector}
\alias{concerto.containsOrderedVector}
\title{concert.containsOrderedVector}
\usage{
concerto.containsOrderedVector(subject, search)
}
\arguments{
\item{subject}{Something}

\item{search}{Something else}
}
\value{
Logical
}
\description{
Undocumented function
}
\author{
Przemyslaw Lis
}

