% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.convertToNumeric.R
\name{concerto.convertToNumeric}
\alias{concerto.convertToNumeric}
\title{concert.convertToNumeric}
\usage{
concerto.convertToNumeric(var)
}
\arguments{
\item{var}{Character vector to be coerced to numeric}
}
\value{
Numeric
}
\description{
Undocumented function
}
\author{
Przemyslaw Lis
}

