% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.db.connect.R
\name{concerto.db.connect}
\alias{concerto.db.connect}
\title{concerto.db.connect}
\usage{
concerto.db.connect(user, password, dbName, host = "localhost", port = 3306,
  dbTimezone)
}
\arguments{
\item{user}{DB User Name}

\item{password}{DB Password for user}

\item{dbName}{Name of DB to connect to}

\item{host}{DB Host Address}

\item{port}{Port number for database}

\item{dbTimezone}{Value used to set time zone parameter of MySQL DB}
}
\description{
Undocumented function
}
\author{
Przemyslaw Lis
}

