% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.initialize.R
\name{concerto.initialize}
\alias{concerto.initialize}
\title{concerto.initialize}
\usage{
concerto.initialize(testID, sessionID, workspaceID, workspacePrefix, user,
  password, dbName, host = "localhost", port = 3306, tempPath, mediaPath,
  dbTimezone, dbConnect, userIP, mediaURL)
}
\arguments{
\item{testID}{ID field for test}

\item{sessionID}{ID field for the testing session}

\item{workspaceID}{ID field for the workspace}

\item{workspacePrefix}{Prefix to use for the workspace}

\item{user}{DB User Name}

\item{password}{DB Password for user}

\item{dbName}{Name of DB to connect to}

\item{host}{DB Host Address}

\item{port}{Port number for database}

\item{tempPath}{File path where temporary files will be stored}

\item{mediaPath}{File path to location where media files are stored}

\item{dbTimezone}{Value used to set time zone parameter of MySQL DB}

\item{dbConnect}{Database connection object}

\item{userIP}{IP address for the user}

\item{mediaURL}{URL specifying where the browser should access media files from}
}
\description{
Undocumented function
}
\author{
Przemyslaw Lis
}

