% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.interpretResponse.R
\name{concerto.interpretResponse}
\alias{concerto.interpretResponse}
\title{concerto.interpretResponse}
\usage{
concerto.interpretResponse()
}
\value{
Response string
}
\description{
Undocumented function
}
\author{
Przemyslaw Lis
}

