% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.qti.contains.R
\name{concerto.qti.contains}
\alias{concerto.qti.contains}
\title{concerto.qti.contains}
\usage{
concerto.qti.contains(exp1, exp2, baseType, cardinality)
}
\arguments{
\item{exp1}{Not clearly documented}

\item{exp2}{Not clearly doocumented}

\item{baseType}{String used to identify if data are paired}

\item{cardinality}{String used to identify if data in exp1 and exp2 are ordered}
}
\description{
Undocumented function
}
\author{
Przemyslaw Lis
}

