% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.qti.delete.R
\name{concerto.qti.delete}
\alias{concerto.qti.delete}
\title{concerto.qti.delete}
\usage{
concerto.qti.delete(exp1, exp2, baseType)
}
\arguments{
\item{exp1}{Not clearly documented}

\item{exp2}{Not clearly documented}

\item{baseType}{Not clearly documented}
}
\value{
Not clearly documented
}
\description{
Not clearly documented
}
\author{
Przemyslaw Lis
}

