% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.qti.equal.R
\name{concerto.qti.equal}
\alias{concerto.qti.equal}
\title{concerto.qti.equal}
\usage{
concerto.qti.equal(arg1, arg2, baseType)
}
\arguments{
\item{arg1}{Not clearly documented}

\item{arg2}{Not clearly documented}

\item{baseType}{Not clearly documented}
}
\value{
Not clearly documented
}
\description{
Not clearly documented
}
\author{
Przemyslaw Lis
}

