% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.qti.get.R
\name{concerto.qti.get}
\alias{concerto.qti.get}
\title{Workspace name retrieval}
\usage{
concerto.qti.get(qtiID, workspaceID = concerto$workspaceID)
}
\arguments{
\item{qtiID}{QTI ID Number}

\item{workspaceID}{Concerto workspace ID.}
}
\value{
Name of the workspace.
}
\description{
Gets workspace name.
}
\examples{
\dontrun{
dbName <- concerto.workspace.get(1)
}
}
\author{
Przemyslaw Lis
}

