% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.qti.initialize.R
\name{concerto.qti.initialize}
\alias{concerto.qti.initialize}
\title{QTI initialization}
\usage{
concerto.qti.initialize(qtiID, params = list(),
  workspaceID = concerto$workspaceID)
}
\arguments{
\item{qtiID}{QTI object ID or name.}

\item{params}{Named list with parameters we want to pass to QTI initialization.}

\item{workspaceID}{Concerto workspace ID where QTI object is located.}
}
\value{
Named list with QTI session state and QTI_HTML/HTML generated by QTI logic
}
\description{
Initializes QTI object.
}
\examples{
\dontrun{
init <- concerto.qti.initialize(1)
}
}
\author{
Przemyslaw Lis
}

