% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.qti.mapResponse.R
\name{concerto.qti.mapResponse}
\alias{concerto.qti.mapResponse}
\title{concerto.qti.mapResponse}
\usage{
concerto.qti.mapResponse(variableName)
}
\arguments{
\item{variableName}{Not clearly documented}
}
\value{
Upper/lower bound of something not documented well
}
\description{
Not clearly documented
}
\author{
Przemyslaw Lis
}

