% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.table.fillSQL.R
\name{concerto.table.fillSQL}
\alias{concerto.table.fillSQL}
\title{Table helper function.}
\usage{
concerto.table.fillSQL(sql, params = list())
}
\arguments{
\item{sql}{SQL code.}

\item{params}{Named list with parameters.}
}
\value{
SQL code with inserted parameters.
}
\description{
Fills SQL code with given parameters.
}
\examples{
\dontrun{
id <- 123
sql <- "SELECT * FROM `db`.`table` WHERE `id`=  id_value  "
sql <- concerto.table.fillSQL(sql, list(id_value = id))
}
}
\author{
Przemyslaw Lis
}

