% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.table.get.R
\name{concerto.table.get}
\alias{concerto.table.get}
\title{Table object retrieval}
\usage{
concerto.table.get(tableID, workspaceID = concerto$workspaceID)
}
\arguments{
\item{tableID}{Table object ID or name.}

\item{workspaceID}{Concerto workspace ID where the table object is located.}
}
\value{
List with table object properties:
\itemize{
\item{id}{table object ID}
\item{name}{table object name}
}
}
\description{
Gets table object.
}
\examples{
\dontrun{
table1 <- concerto.table.get(1)
table2 <- concerto.table.get("some_other_table")
}
}
\author{
Przemyslaw Lis
}

