% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.table.query.R
\name{concerto.table.query}
\alias{concerto.table.query}
\title{Table query}
\usage{
concerto.table.query(sql, params = list())
}
\arguments{
\item{sql}{SQL query to perform.}

\item{params}{Named list with parameters.}
}
\value{
Returns result set.
}
\description{
Perform query on Concerto table.
}
\details{
Contains extended function wizard with SQL query builder when used in Concerto platform.
}
\author{
Przemyslaw Lis
}

