% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.template.loader.R
\name{concerto.template.loader}
\alias{concerto.template.loader}
\title{Loader template}
\usage{
concerto.template.loader(templateID = -1, HTML = "", head = "",
  params = list(), workspaceID = concerto$workspaceID,
  effectShow = "default", effectShowOptions = "default",
  effectHide = "default", effectHideOptions = "default")
}
\arguments{
\item{templateID}{Template ID or name.}

\item{HTML}{Template HTML.}

\item{head}{Template head tag.}

\item{params}{Named list with parameters we want to be inserted into template HTML.}

\item{workspaceID}{Concerto workspace ID where template object is located.}

\item{effectShow}{Template show transition effect.}

\item{effectShowOptions}{Template show transition effect options list.}

\item{effectHide}{Template hide transition effect.}

\item{effectHideOptions}{Template hide transition effect options list.}
}
\description{
Sets a template that will be shown when the proper templates are loaded.
}
\examples{
\dontrun{
response1 <- concerto.template.loader(1)
response2 <- concerto.template.loader(HTML='<div align='center'>loading ... please wait ...</div>')
}
}
\author{
Przemyslaw Lis
}

