% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.test.getReturnVariables.R
\name{concerto.test.getReturnVariables}
\alias{concerto.test.getReturnVariables}
\title{concerto.test.getReturnVariables}
\usage{
concerto.test.getReturnVariables(testID, workspaceID = concerto$workspaceID)
}
\arguments{
\item{testID}{ID for the assessment}

\item{workspaceID}{ID for the workspace}
}
\description{
Accessor method for return variables
}
\author{
Przemyslaw Lis
}

