% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.test.updateAllReturnVariables.R
\name{concerto.test.updateAllReturnVariables}
\alias{concerto.test.updateAllReturnVariables}
\title{concerto.test.updateAllReturnVariables}
\usage{
concerto.test.updateAllReturnVariables()
}
\description{
Not clearly documented
}
\author{
Przemyslaw Lis
}

