% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.test.updateReturnVariable.R
\name{concerto.test.updateReturnVariable}
\alias{concerto.test.updateReturnVariable}
\title{concerto.test.updateReturnVariable}
\usage{
concerto.test.updateReturnVariable(variable)
}
\arguments{
\item{variable}{Not clearly documented}
}
\description{
Not clearly documented
}
\author{
Przemyslaw Lis
}

