% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateHead.R
\name{concerto.updateHead}
\alias{concerto.updateHead}
\title{concerto.updateHead}
\usage{
concerto.updateHead(html)
}
\arguments{
\item{html}{Not clearly documented}
}
\description{
Not clearly documented
}
\author{
Przemyslaw Lis
}

