% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateHTML.R
\name{concerto.updateHTML}
\alias{concerto.updateHTML}
\title{concerto.updateHTML}
\usage{
concerto.updateHTML(html)
}
\arguments{
\item{html}{Not clearly documented}
}
\description{
Not clearly documented
}

