% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateRelease.R
\name{concerto.updateRelease}
\alias{concerto.updateRelease}
\title{concerto.updateRelease}
\usage{
concerto.updateRelease(release)
}
\arguments{
\item{release}{Not clearly documented}
}
\description{
Not clearly documented
}
\author{
Przemyslaw Lis
}

