% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateStatus.R
\name{concerto.updateStatus}
\alias{concerto.updateStatus}
\title{concerto.updateStatus}
\usage{
concerto.updateStatus(status)
}
\arguments{
\item{status}{Test status}
}
\description{
Set method for status object
}
\author{
Przemyslaw Lis
}

