% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateTemplateID.R
\name{concerto.updateTemplateID}
\alias{concerto.updateTemplateID}
\title{concerto.updateTemplateID}
\usage{
concerto.updateTemplateID(templateID)
}
\arguments{
\item{templateID}{ID field for the assessment template to use}
}
\description{
Set method for template ID object
}
\author{
Przemyslaw Lis
}

