% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateTemplateWorkspaceID.R
\name{concerto.updateTemplateWorkspaceID}
\alias{concerto.updateTemplateWorkspaceID}
\title{concerto.updateTemplateWorkspaceID}
\usage{
concerto.updateTemplateWorkspaceID(workspaceID)
}
\arguments{
\item{workspaceID}{ID field for the workspace}
}
\description{
Set method for workspace ID object
}
\author{
Przemyslaw Lis
}

