% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concerto.updateTimeLimit.R
\name{concerto.updateTimeLimit}
\alias{concerto.updateTimeLimit}
\title{concerto.updateTimeLimit}
\usage{
concerto.updateTimeLimit(timeLimit)
}
\arguments{
\item{timeLimit}{value used to set the timelimit}
}
\description{
Undocumented function
}
\author{
Przemyslaw Lis
}

