\name{concerto.workspace.get}
\alias{concerto.workspace.get}
\title{
Workspace name retrieval
}
\description{
Gets workspace name.
}
\usage{
concerto.workspace.get(workspaceID)
}
\arguments{
  \item{workspaceID}{
Concerto workspace ID.
}
}
\details{
}
\value{
Name of the workspace.
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
dbName <- concerto.workspace.get(1)
}
\keyword{ workspace }
\keyword{ MySQL }
\keyword{ database }
